Verma: products are now materialised from smallest to largest
This monster of a commit implements the long-overdue refactoring of #27 while attempting to fix a bug introduced by #24, replacing most support structures with inlined attributes in the edges of the automata. This enormously simplifies most of the implementation.

After this commit, the only retained support structure is the one holding the solver variables associated with each transition. I debate myself whether to fold them into the transition types in the same fashion.

Closes: #24, #27